Rename file containing handlers for station-specific resources
