+++ index.js
@@ -7,7 +7,7 @@ const express = require('express')
 const _ = require('lodash')
 const processenv = require('processenv')
 // const hfc = require('./handlers/forecast')
-const hda = require('./handlers/data_access')
+const sh = require('./handlers/stations')
 // const hm = require('./handlers/measurements')
 // const hpoi = require('./handlers/poi')
 // const poifce = require('./lib/poi_forecast_engine')
@@ -283,15 +283,15 @@ async function init () {
   })
 
   // Define routing
-  backend.register('getFilteredListOfStations', hda.getWeatherStations(stationCatalog))
-  backend.register('getStation', hda.getSingleWeatherStation(stationCatalog))
+  backend.register('getFilteredListOfStations', sh.getWeatherStations(stationCatalog))
+  backend.register('getStation', sh.getSingleWeatherStation(stationCatalog))
   backend.register(
     'getMeasuredValues',
-    hda.getMeasuredValues(DATA_ROOT_PATH, voisDataAccessConfigs)
+    sh.getMeasuredValues(DATA_ROOT_PATH, voisDataAccessConfigs)
   )
   backend.register(
     'getForecastAtStation',
-    hda.getForecastAtStation(DATA_ROOT_PATH, voisDataAccessConfigs, stationCatalog)
+    sh.getForecastAtStation(DATA_ROOT_PATH, voisDataAccessConfigs, stationCatalog)
   )
 
   // Handle unsuccessful requests
